Continue adding functionality to the annotation app.
One can now see the selected image in the axis.
One can now select squares in the image.
One can now save the selected squares.
* annotation/annotation_gui.m (file_list_Callback) The image is
refreshed every time that a file is selected on the file list.
(save_annotation_Callback) The callback calls the function
save_annotation that is implemented in the annotation/save_annotation.m
file.
(labels_Callback) Added variable that will keep track of the label that
is currently chosen.  The created square will be related to the
currently selected label.
(add_files_Callback) Function executed when we push the 'Add Files'
button.  It calls a Matlab predefined thingy that returns a file list
and a directory. (easy peazy :)
(put_image_in_axis) Reads the image from the file system and places it
in the axis.
(button_pressed_on_image) This callback I had to manually create to
detect a mouse button press on the axis.  This is the function that
contains all the logic to create the square and paint it and ....
(annotation_init) New helper function to create a empty annotation
structure.
(read_annotations) New function to read the annotations.  I still have
to implement it.
(drawbox) Helper function to draw a red box. Was implemented by PASCAL.
* annotation/save_annotation.m: New file containing the save_annotation
function.  Was implemented by PASCAL.