Implemented the read annotation functionality.
* annotation/annotation_gui.m (file_list_Callback) Call the new function
put_annotations_in_axis to put the read regions into the new image.
(button_pressed_on_image) Decided to create a new fule called drawbox.
This is necessary because that function is called from two different
files.
(annotation_init) This is going to a new file because it is needed from
two different files.
(read_annotations):
* annotation/drawbox (drawbox): New file containing the drawbox function
that draws a red box in the specified regions.  This comes from
annotation/annotation_gui.m.
* annotation/annotation_init.m: Took this function out of
annotation/annotation_gui.m
* annotation/put_annotations_in_axis.m (put_annotation_in_axis) : New
function that traverses the annotation region and paints each region on
the current axis.