Implement the change image functionality in a different function.
Matlab seems to not want to keep the handles data created in the new
function.  This commit has a semy-hack to avoid strange behavior.
* annotation/annotation_gui.m (file_list_Callback)
(add_files_Callback, on_key_press_callback): Call the new function.
(select_offset_from_list): This is a new function that gathered the
funcitionality of selecting a file from list.