Implement the clear regions button.
* annotation/annotation_gui.m (clear_annotation_Callback): We traverse
all the regions in the curr_ann and set the handler Visible property to
'off' of the texts and lines.  Once everything is hidden, we create a
new curr_ann.  We do not reload the image.
(button_pressed_on_image, select_offset_from_list):
* annotation/annotation_init.m (annotation_init): Add an element that
holds the text and line handlers.
* annotation/drawbox.m (drawbox): Return the text and line handlers.
* annotation/put_annotations_in_axis.m (put_annotations_in_axis): The
function now returns a new curr_ann object.  This is an ongoing hack
that I have to use so the information gets passed correctly.  Matlab,
you win again...