Added a restreach functionality.
It is now possible to re-streach the last annotation box that is created.
I had to rework the way the regions in the curr_ann struct where
handled.  The last regions is now pointed at by a regions offset called
reg_offset.  I created two new variables in the region struct: a active
var, to signify that that regions is active or not (Will come in handy
when implementing the delete functionality) and a bbox_figure var that
holds the information for the rbbox with respect to the figure as
opposed to the axis.

* annotation/annotation_gui.m (annotation_gui_OpeningFcn): Changed the
way we initialize the curr_ann.regions var.  We now initialize the
counter (reg_offset) to 0 and initialize the first regions element to
something that is inactive.
(file_list_Callback):
(clear_annotation_Callback): This function resets the reg_offset to 0
and resets the first element in the regions vector.  It also erases all
the squares from the current figure, image.  The regions vector is not
erased.
(button_pressed_on_image): This is where the new re-stretching is
implemented.  It only modifies the latest created box.  It works with
the right click of the mouse and only functions after a creation of a
box.  It supports re-stretching from all four sides of the annotation
square.
* annotation/annotation_init.m: Added a couple of new vars to the init
scripts.  It now includes bbox_figure and reg_offset with the other
vars.
* annotation/read_annotation.m: We reworked the construction of the
curr_ann from this function.  It now creates all the necessary elements
that curr_ann needs.  Including the active flag and the bbox_figure
vector.  Since we don't have any info to create the bbox_figure vector
from the file, we will initialize it to [].
* annotation/save_annotation.m: We reworked the saving process in such a
way that only the active regions get saved (the inactive are ignored).
* annotation/put_annotation_in_axis.m: Changed the process of creating
the annotations for the axis to understand the new standard data
structure.  It now looks for the active flag and only paints the active
members.