Implement the correction feature.
This allows the user to select a region and delete it.
* annotation/annotation_gui.m (button_pressed_on_image): Make sure we
don't do anything when handles.correction.active is '1'.  The onclick
callback for the image should be un-responsive.
(button_press_on_line): New function.  It will be called when the user
left clicks on a pixel that is close to a line.  This will cause a new
box to be created on top of the existing one.  Various internal
variables get set to hole the green square temporarily.
(on_key_press_callback):  Add a functionality for when th user presses
'd' or 'D'.  If the correct.active state is '1', this action will delete
the selected box.  The selected box is represented by variables held in
handles.correction.
(annotation_gui): Create new variables to hold the selected box when in
correction mode.  We need to specify where the box to be corrected is
located. The handles for the text and the lines and the coordinates.  We
also hold the correction state.  Its off by default.
(correction_toggle_callback): This function toggles the correction
state.
* annotation/drawbox.m (drawbox): This function now receives a new
argument. The callback function for the onclick action on the line.
When it draws the line it also sets the callback.
* annotation/put_annotations_in_axis.m (put_annotations_in_axis):
This function now accepts an additional argument.  The callback
function for the onclick of the line.  This, in turn, is passed to
the drawbox function.
* annotation/annotation_gui.fig: Create a new button to toggle the
correction state.