Make sure we handle the ftp errors correctly
For our purposes we will consider a successful ftp connection when the
server, username, password and directory are all valid.  If one of these
is missing we will error out.

* annotation/ftp_conf_gui.m: Add the checks for ftp connection and for
directory validity.  We used matlabs try-exception.
* annotation/ftp_connection.m: The auto-generated code from Matlab's
"guide" contains a function that does not accept a "empty" char as
argument.  Work around that issue by checking for empty chars.