Add the underlying ssh code.
All ssh commands work with the unix based 'ssh' and 'scp' commands.  I
used the 'unix' command in Matlab.
* annotation/annotation.conf: We added ssh specific configuration
variables: ssh_server, ssh_username, ssh_dir.
* annotation/annotation_conf.m: Added code to read the new ssh
configuration variables.
* annotation/ssh_delete.m: New function that deletes a file on the
server
* annotation/ssh_exists.m: ascertains if a file exists or not on the
server.
* annotation/ssh_download.m: Downloads a file from the server to the
client.
* annotation/ssh_getfile.m: Downloads a file from the server to the
client and takes additional care of locks and other users.
* annotation/ssh_getlist.m: Downloads the file list of the current ssh
directory.  The ssh directory is defined in the ssh_struct structure.
* annotation/ssh_lck.m: Has four functions. 1. lock, 2. unlock, 3.
islocked and 4. ismine.  The islocked function returns whether a file is
locked or not.  The ismine function returns whether the user owns the
lock or not.
* annotation/ssh_sane.m: Returns 1 if the ssh connection is good.  Note
that for the ssh functionality to work the user must configure
passwordless ssh access from the client to the server.  Something around
the lines of http://www.debian-administration.org/articles/152
* annotation/ssh_savefile.m: Save a file from the client to the server.
It takes care of locks and other users.
* annotation/ssh_upload.m: Saves a file from the client to the server.