Introduce the ssh code into annotation_gui.m
We are trying to generalize the way the calls are made by introducing
intermidiate files like annotation_getfile and annotation_settype.
* annotation/annotation_getfile.m: Intermediate file that decides to use
ssh or ftp depending on the name string.
* annotation/annotation_settype.m: Intermediate file that modifies the
annotation based on the file name.  It modifies the ssh and ftp flags.
* annotation/ssh_download.m: Reduced the way we check for ssh sanity.
* annotation/ssh_exists.m: Likewise.
* annotation/ssh_getlist.m: Likewise.
* annotation/ssh_delete.m: Likewise.
* annotation/ssh_upload.m: Likewise.
* annotation/ssh_sane.m: Introduced the error messages in the sane check
to reduce the size in the other files.
* annotation/annotation_gui.fig : Add an ssh button.
* annotation/annotation_gui.m : Start using the new intermediate files.