Only use im* in the creation and correction
The im* functions are too CPU intensive.  We will only use them when the user is
creating and modifying the annotations.  The annotations can now be removed with
the Esc key.
* src/annotation_gui.m (file_list_Callback): Check to see if we are not in
isModifying.
(button_pressed_on_image): Use impoly to create the annotation.  But delete it
as soon as the user is done.  In its stead put a series of line objects.  The
annotation label did not really change.
(button_press_on_line):  New callback function.  Activated after pressing on the
line.  We erase the pressed line and put a impoly in its stead.  This will allow
the user to modify the annotation.  Esc removes annotation.
(button_pressed_on_text_label): Removed annotation removal logic.  It is now in
the button_press_on_line function.
(select_offset_from_list): Check to see if we are not in isModifying.
(remove_Callback): Removed the 'remove' button and its callback.
* src/annotation_gui.fig: Likewise.
* src/annotation_init.m (annotation_init):  Introduced region.rect and
region.line_handle.  The first contains the coordinates of the surrounding rectangle.
The second contains the line handle.
* src/annotation_read.m (annotation_readVer20): Use the new rect and line_handle
variables.
* src/annotation_save.m (annotation_save_VER20): Likewise.