Introduce a table of Elements.
Elements will represent different annotations that are "the same" between
images.
* src/pyann/annData.py (initDB): Add new table to the creation script.
(addElement): New method to add an element.
(addAnnotation): Change the way we accept arguments in this method.
(initAnn): New method that adds an annotation with a new element row.
(appendAnn): New method that adds an annotation and relates it to an existing
element.
* tests/db_test.py (test_annotation, test_picturesByDate)
(test_picturesByPlotID): Change tests respectfully.