Activate the filling of the file list
This is painful. There are basically two ways of 'uploading' files: using
input type=file and using the drag and drop facility. The drag and drop is
overkill for this application and the input type=file cannot be styled. I
resolved to create a very ugly hack where the un-styled input is used in stead
of the original pretty button. GRRRR!!!!!!
References:
* http://stackoverflow.com/questions/2855589/replace-input-type-file-by-an-image
* http://stackoverflow.com/questions/210643/in-javascript-can-i-make-a-click-event-fire-programmatically-for-a-file-input

* src/html5ann/ann.html: Create and style input element.
* src/html5ann/ann.js (ann_input.onchange, zoom): Use the onchange callback to
put files into the filelist.
* src/html5ann/ann_init.js (ann_list) create a js var that
contains the filelist select.