Found a way to avoid ugly hack !!!
One can create a function with some predifined variable elements (not the
arguments). This allows me to define the 'this' element on an 'onclick'
function before the onclick is executed. When onclick is executed, the 'this'
is not Window, its whatever I defined.
* src/html5ann/ann_filelist.js (FileList, FileList.append_files)
(FileList.remove_file): Remove Ugly hack. Long live cool hacks!!!!