Put the csv annotations in the image.
* src/html5ann/ann_canvas.js (AnnCanvas): Put an initial image on the svg
when we begin.
(csvOnCanvas): New method to add polygons to svg canvas.
(imgOnSVG): New method to add an image to svg canvas. We don't create a new
svg container every time. We just change the href attribute.
(clsSVG): New method to clean the svg canvas DOM element. Its tricky because
we need to clean all but the first three elements. First two element are
created automatically by Raphael (soon to be removed) and the last is the
image.
* src/html5ann/ann_filelist.js: Use newly created methods in ann_canvas.js.
* src/html5ann/ann.html: Source the new file ann_csv.js
* src/html5ann/ann_csv.js: New file in charge of parsing the CSV files.