Add the this.actions variable
This variable will keep track in what state of action we are in. We can be in
panzoom(default), polygons and rectangles. Each mouse action (mousedown,
mouseup, mousemove and mousescroll) will have to have logic for each action
state. For now the polygon and rectangle action states have no particular
difference from the panzoom other than panning and zooming is disabled.