Implement one click annotation.
* src/matann/annotation_gui.m (button_pressed_on_image): Just create a line
representing a square. We don't use imrect.
(button_press_on_line): Reconstruct the square with the line leading the the
label. Label is placed using handles.roiDistConst. Always be a rect when user
wants to modify annotation.
(on_move_roi): Be sure to put the label were it is expected.